refactor: 폴더 및 파일명 변경, 주석 제거
