+++ src/static/category.ts
@@ -9,6 +9,6 @@ export const SUB_CATEGORY = [
 export const MAIN_CATEGORY = [
   { id: 1, name: 'Home', path: '/' },
   { id: 2, name: 'Normal', path: '/basicauction' },
-  // { id: 3, name: 'Live', path: '/liveauction' },
+  { id: 3, name: 'Live', path: '/liveauction' },
   { id: 4, name: 'How To', path: '/howto' },
 ];
